Move bundler files in their own folders
