+++ src/__tests__/webpack/modules.test.ts
@@ -2,10 +2,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { LocalModule } from '../types';
+import { LocalModule } from '../../types';
 
 describe('Modules', () => {
-    const { Modules } = require('../modules');
+    const { Modules } = require('../../webpack/modules');
     // Webpack5 is actually throwing an error when using this property.
     const getThrowingDependency = (dep: any) => {
         Object.defineProperty(dep, 'module', {
